Split errors and protocol errors test
