+++ tests/test_errors.py
@@ -3,44 +3,23 @@ from __future__ import annotations
 import asyncio
 import socket
 import sys
-from typing import TYPE_CHECKING, TypeVar, cast
+from typing import TYPE_CHECKING, cast
 
 import pytest
 
-from pyqwest import (
-    Client,
-    HTTPTransport,
-    ReadError,
-    RemoteProtocolError,
-    SyncClient,
-    SyncHTTPTransport,
-    WriteError,
-)
+from pyqwest import Client, ReadError, SyncClient, WriteError
 
-from ._util import (
-    BAD_CHUNKED_FRAMING,
-    GARBAGE_RESPONSE,
-    MALFORMED_STATUS_LINE,
-    NO_RESPONSE,
-    TRUNCATED_CHUNKED_RESPONSE,
-    TRUNCATED_RESPONSE,
-    SyncRequestBody,
-    raw_server,
-)
+from ._util import SyncRequestBody
 
 if TYPE_CHECKING:
-    from collections.abc import AsyncIterator, Callable, Iterator
+    from collections.abc import AsyncIterator, Iterator
     from queue import Queue
 
-_Test = TypeVar("_Test", bound="Callable[..., object]")
 
-
-# Applied per test rather than as a module-level pytestmark because the protocol
-# error tests at the bottom of this file serve their own raw socket and so do not
-# take the shared server fixtures these parametrize.
-def uses_h2_server(fn: _Test) -> _Test:
-    fn = pytest.mark.parametrize("http_version", ["h2"], indirect=True)(fn)
-    return pytest.mark.parametrize("http_scheme", ["http"], indirect=True)(fn)
+pytestmark = [
+    pytest.mark.parametrize("http_scheme", ["http"], indirect=True),
+    pytest.mark.parametrize("http_version", ["h2"], indirect=True),
+]
 
 
 async def request_body(queue: asyncio.Queue) -> AsyncIterator[bytes]:
@@ -60,7 +39,6 @@ def sync_request_body(queue: Queue) -> Iterator[bytes]:
 
 
 @pytest.mark.asyncio
-@uses_h2_server
 async def test_request_timeout(client: Client | SyncClient, url: str) -> None:
     if sys.version_info < (3, 11):
         pytest.skip("asyncio.timeout requires Python 3.11+")
@@ -93,7 +71,6 @@ async def test_request_timeout(client: Client | SyncClient, url: str) -> None:
 
 
 @pytest.mark.asyncio
-@uses_h2_server
 async def test_response_content_timeout(client: Client | SyncClient, url: str) -> None:
     if sys.version_info < (3, 11):
         pytest.skip("asyncio.timeout requires Python 3.11+")
@@ -125,7 +102,6 @@ async def test_response_content_timeout(client: Client | SyncClient, url: str) -
 
 
 @pytest.mark.asyncio
-@uses_h2_server
 async def test_connection_error(
     client: Client | SyncClient, client_type: str, url: str
 ) -> None:
@@ -150,7 +126,6 @@ async def test_connection_error(
 
 
 @pytest.mark.asyncio
-@uses_h2_server
 async def test_request_not_bytes(client: Client | SyncClient, url: str) -> None:
     method = "POST"
     url = f"{url}/echo"
@@ -173,65 +148,3 @@ async def test_request_not_bytes(client: Client | SyncClient, url: str) -> None:
 
             async with client.stream(method, url, content=request_content()) as resp:
                 await anext(resp.content)
-
-
-# Each of these is a way for the peer to break HTTP framing, and each is served
-# over a raw socket so the fault is exactly what the test names. They are
-# HTTP/1.1 by construction, so unlike the tests above they build their own
-# transport rather than using the parametrized fixtures.
-PROTOCOL_FAULTS = [
-    pytest.param(NO_RESPONSE, id="no-response"),
-    pytest.param(MALFORMED_STATUS_LINE, id="malformed-status-line"),
-    pytest.param(GARBAGE_RESPONSE, id="garbage-response"),
-    pytest.param(BAD_CHUNKED_FRAMING, id="bad-chunked-framing"),
-    pytest.param(TRUNCATED_RESPONSE, id="truncated-body"),
-    pytest.param(TRUNCATED_CHUNKED_RESPONSE, id="truncated-chunked-body"),
-]
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("response", PROTOCOL_FAULTS)
-async def test_async_protocol_error(response: bytes) -> None:
-    with raw_server(response) as url:
-        async with HTTPTransport() as transport:
-            with pytest.raises(RemoteProtocolError):
-                await Client(transport).get(url)
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("response", PROTOCOL_FAULTS)
-async def test_sync_protocol_error(response: bytes) -> None:
-    def run() -> None:
-        with (
-            raw_server(response) as url,
-            SyncHTTPTransport() as transport,
-            pytest.raises(RemoteProtocolError),
-        ):
-            SyncClient(transport).get(url)
-
-    await asyncio.to_thread(run)
-
-
-@pytest.mark.asyncio
-async def test_async_response_reset_is_read_error() -> None:
-    # A response cut short by a reset is a broken connection rather than a
-    # protocol violation, so it must not be swept up by the tests above.
-    with raw_server(TRUNCATED_RESPONSE, reset=True) as url:
-        async with HTTPTransport() as transport:
-            # Usually a read error but timing can make it a write error,
-            # especially on Windows.
-            with pytest.raises((ReadError, WriteError)):
-                await Client(transport).get(url)
-
-
-@pytest.mark.asyncio
-async def test_sync_response_reset_is_read_error() -> None:
-    def run() -> None:
-        with (
-            raw_server(TRUNCATED_RESPONSE, reset=True) as url,
-            SyncHTTPTransport() as transport,
-            pytest.raises((ReadError, WriteError)),
-        ):
-            SyncClient(transport).get(url)
-
-    await asyncio.to_thread(run)
